pyproject.toml
[project]
name = "gem-read"
version = "1.0.0"
description = "Gem Read is an AI-assisted PDF/EPUB reader with Gemini integration"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "PySide6>=6.7",
    "PyMuPDF>=1.24",
    "google-genai>=1.0",
    "qasync>=0.27",
    "python-dotenv>=1.0",
    "Pillow>=10.0",
    "platformdirs>=4.0",
    "markdown>=3.5",
    "uvicorn>=0.44.0",
    "fastapi>=0.135.3",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pdf_epub_reader", "src/browser_api"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
